10.2-Redux Перенес initial states для элементов в InitialStates.ts из Constants.ts
