+++ src/Model/Store/InitialStates.ts
@@ -2,7 +2,13 @@ import {Editor} from "../Types/Editor";
 import {Presentation} from "../Types/Presentation";
 import {Slide} from "../Types/Slide";
 import {generate} from "../../Utils/generate";
-import {DEFAULT_PRESENTATION_TITLE, DEFAULT_SLIDE_BACKGROUND_COLOR, DEFAULT_SLIDE_TITLE} from "../Constants";
+import {
+    DEFAULT_ELEMENT_FILL,
+    DEFAULT_PRESENTATION_TITLE,
+    DEFAULT_SLIDE_BACKGROUND_COLOR,
+    DEFAULT_SLIDE_TITLE, DEFAULT_TEXT_COLOR
+} from "../Constants";
+import {Circle, CommonType, Picture, Rectangle, TextBox, Triangle} from "../Types/Element";
 
 export const getInitialSlideState = (): Slide => {
     return {
@@ -13,14 +19,42 @@ export const getInitialSlideState = (): Slide => {
     }
 }
 
-export const initialPresentationState: Presentation = {
+export const INITIAL_PRESENTATION_STATE: Presentation = {
     title: DEFAULT_PRESENTATION_TITLE,
     slides: [getInitialSlideState()]
 }
 
-export const initialEditorState: Editor = {
-    presentation: initialPresentationState,
-    selectedSlideIds: [initialPresentationState.slides[0].id],
+export const INITIAL_EDITOR_STATE: Editor = {
+    presentation: INITIAL_PRESENTATION_STATE,
+    selectedSlideIds: [INITIAL_PRESENTATION_STATE.slides[0].id],
     selectedElementIds: [],
-    currentSlideId: initialPresentationState.slides[0].id
+    currentSlideId: INITIAL_PRESENTATION_STATE.slides[0].id
+}
+
+const INITIAL_COMMON_STATE: CommonType = {
+    border: null,
+    fill: DEFAULT_ELEMENT_FILL,
+    height: 200,
+    width: 200,
+    x: 0,
+    y: 0
+}
+
+export const INITIAL_RECTANGLE_STATE: Rectangle = INITIAL_COMMON_STATE;
+
+export const INITIAL_TRIANGLE_STATE: Triangle = INITIAL_COMMON_STATE;
+
+export const INITIAL_CIRCLE_STATE: Circle = INITIAL_COMMON_STATE;
+
+export const INITIAL_PICTURE_STATE: Picture = {
+    ...INITIAL_COMMON_STATE,
+    src: 'url(\'\')'
+}
+
+export const INITIAL_TEXTBOX_STATE: TextBox = {
+    ...INITIAL_COMMON_STATE,
+    content: "Text...",
+    fontFamily: "",
+    fontSize: 0,
+    textColor: DEFAULT_TEXT_COLOR,
 }
